Port auto-struct mapping to core relation
