+++ core/spec/unit/rom/relation/call_spec.rb
@@ -2,7 +2,11 @@ require 'rom/relation'
 
 RSpec.describe ROM::Relation, '#call' do
   subject(:relation) do
-    relation_class.new(data)
+    relation_class.new(data, options)
+  end
+
+  let(:options) do
+    {}
   end
 
   context 'without read types in schema' do
@@ -46,6 +50,35 @@ RSpec.describe ROM::Relation, '#call' do
     it 'returns loaded relation with coerced data' do
       expect(relation.call.collection).
         to eql([{ id: 1, name: 'Jane' }, { id: 2, name: 'John'} ])
+    end
+  end
+
+  describe 'auto-struct' do
+    let(:relation_class) do
+      Class.new(ROM::Relation[:memory]) do
+        schema(:users) do
+          attribute :id, ROM::Types::Int
+          attribute :name, ROM::Types::String
+        end
+      end
+    end
+
+    let(:options) do
+      { auto_struct: true }
+    end
+
+    let(:data) do
+      [{ id: 1, name: 'Jane' }, { id: 2, name: 'John'} ]
+    end
+
+    it 'automatically maps to structs' do
+      result = relation.call.to_a
+
+      expect(result[0].id).to be(1)
+      expect(result[0].name).to eql('Jane')
+
+      expect(result[1].id).to be(2)
+      expect(result[1].name).to eql('John')
     end
   end
 end
